Update UnoTFT_DS18B20.ino
Changed the disconnect "temp" to 11.11 so it is easier to detect.
Added a 6-bit array named errorBits which track disconnects.  That array is printed to the screen as a binary number to show which sensor has errored out.  A bit is cleared when that sensor returns a temperature change reading.